% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgcri.colors.R
\name{jgcri.colors}
\alias{jgcri.colors}
\title{Colors Schemes used in the figsjgcri package}
\usage{
jgcri.colors()
}
\description{
Colors Schemes used in the figsjgcri package
}
\examples{
jgcri.colors()
}
\keyword{1,}
\keyword{2}
\keyword{keyword}
